+++ src/netorcai/metaprotocol_version.d
@@ -1,0 +1,11 @@
+module netorcai.metaprotocol_version;
+
+int majorVersion = 2;
+int minorVersion = 0;
+int patchVersion = 0;
+
+string metaprotocolVersion()
+{
+    import std.format;
+    return format!"%d.%d.%d"(majorVersion, minorVersion, patchVersion);
+}